aider: Added tooltip to the agent's name in `editors/rift-vscode-svelte/webviews/components/logs/Agent.svelte`.
# Context:
USER: in `editors/rift-vscode-svelte/webviews/components/logs/Agent.svelte` let's add a tooltip that appears on hovering over the name of the agent.
ASSISTANT: To add a tooltip that appears on hovering over the name of the agent, we need to modify the `editors/rift-vscode-svelte/webviews/components/logs/Agent.svelte` file.

We will add a `title` attribute to the HTML element that contains the agent's name. The `title` attribute is used to specify extra information about an element and the content is shown as a tooltip when the mouse hovers over the element.

Here is the change:

```svelte
editors/rift-vscode-svelte/webviews/components/logs/Agent.svelte
<<<<<<< ORIGINAL
            {name}
=======
            <span title="This is the agent's name">{name}</span>
>>>>>>> UPDATED
```

This will show a tooltip with the text "This is the agent's name" when the mouse hovers over the agent's name. You can replace the text with any description you want to show in the tooltip.