aider: Change the `rift.mockTypeCommandP` command to show a list of all files in the workspace when executed.
# Aider chat conversation:

USER: is there any way I can edit the rift.mockTypeCommandP file to show me a preview of all the files I can open?

ASSISTANT: The `rift.mockTypeCommandP` is not a file, but a command registered in the `extension.ts` file. This command currently allows you to type a file name and then shows you a list of files that match the name you typed. If you want to modify this command to show a preview of all files in the workspace, you can do so by modifying the `extension.ts` file.

Here's how you can do it:

```typescript
editors/rift-vscode/src/extension.ts
<<<<<<< ORIGINAL
let disposableNuke = vscode.commands.registerCommand(
  'rift.mockTypeCommandP',
  async () => {
      const input = await vscode.window.showInputBox({
          prompt: 'Type a file name',
      });

      if (input) {
        console.log('input')
      const files = await vscode.workspace.findFiles(`**/${input}`);

          if (files && files.length > 0) {
              const filePaths = files.map(file => file.fsPath);
          vscode.window.showQuickPick(filePaths).then(selectedPath => {
              if (selectedPath) {
                  // Open the selected file in a new editor
                  vscode.workspace.openTextDocument(selectedPath).then(doc => {
                      vscode.window.showTextDocument(doc, { preview: false });
                  });
              }
          });
          }
      }
  }
);
=======
let disposableNuke = vscode.commands.registerCommand(
  'rift.mockTypeCommandP',
  async () => {
      const files = await vscode.workspace.findFiles('**/*');

      if (files && files.length > 0) {
          const filePaths = files.map(file => file.fsPath);
          vscode.window.showQuickPick(filePaths).then(selectedPath => {
              if (selectedPath) {
                  // Open the selected file in a new editor
                  vscode.workspace.openTextDocument(selectedPath).then(doc => {
                      vscode.window.showTextDocument(doc, { preview: false });
                  });
              }
          });
      }
  }
);
>>>>>>> UPDATED
```

This change will make the `rift.mockTypeCommandP` command show a list of all files in the workspace when it is executed. You can then select a file from the list to open it.